+++ packages/terra-core-docs/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 1.52.0 - (November 21, 2023)
+
 * Added
   * Added test to `terra-section-header` for the fixed section header title.